video_stream_service/requirements.txt
anyio==3.7.0
click==8.1.3
dlib==19.24.2
exceptiongroup==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.96.0
h11==0.14.0
hazelcast-python-client==5.2.0
idna==3.4
kafka-python==2.0.2
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pydantic==1.10.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
